refactor: remove confirmed-dead code (~2.6k LOC)
Delete code with zero non-test callers (grep-verified): the unused Claim/Proof
pipelines (lifecycle handles submission) + DefaultProofRequirementChecker;
supplier_drain.go (the drain manager is never constructed); SessionWindow /
CalculateSessionWindow (callers use sharedtypes helpers); 19 superseded KeyBuilder
methods + dead ParamsSession*/MeterLimit* keys; a commented-out calculateFee dup.
Shared test mocks extracted to mock_clients_test.go. UpdateWALPosition left for a
follow-up (coordinated interface edit).
