+++ miner/session_lifecycle.go
@@ -1304,51 +1304,3 @@ func (m *SessionLifecycleManager) Close() error {
 	m.logger.Info().Msg("session lifecycle manager closed")
 	return nil
 }
-
-// SessionWindow represents the timing windows for a session.
-type SessionWindow struct {
-	SessionEndHeight int64
-	GracePeriodEnd   int64
-	ClaimWindowOpen  int64
-	ClaimWindowClose int64
-	ProofWindowOpen  int64
-	ProofWindowClose int64
-}
-
-// CalculateSessionWindow calculates all timing windows for a session.
-func CalculateSessionWindow(params *sharedtypes.Params, sessionEndHeight int64) SessionWindow {
-	return SessionWindow{
-		SessionEndHeight: sessionEndHeight,
-		GracePeriodEnd:   sessionEndHeight + int64(params.GetGracePeriodEndOffsetBlocks()),
-		ClaimWindowOpen:  sharedtypes.GetClaimWindowOpenHeight(params, sessionEndHeight),
-		ClaimWindowClose: sharedtypes.GetClaimWindowCloseHeight(params, sessionEndHeight),
-		ProofWindowOpen:  sharedtypes.GetProofWindowOpenHeight(params, sessionEndHeight),
-		ProofWindowClose: sharedtypes.GetProofWindowCloseHeight(params, sessionEndHeight),
-	}
-}
-
-// IsInClaimWindow returns true if the current height is within the claim window.
-func (w SessionWindow) IsInClaimWindow(currentHeight int64) bool {
-	return currentHeight >= w.ClaimWindowOpen && currentHeight < w.ClaimWindowClose
-}
-
-// IsInProofWindow returns true if the current height is within the proof window.
-func (w SessionWindow) IsInProofWindow(currentHeight int64) bool {
-	return currentHeight >= w.ProofWindowOpen && currentHeight < w.ProofWindowClose
-}
-
-// BlocksUntilClaimWindowClose returns blocks remaining until claim window closes.
-func (w SessionWindow) BlocksUntilClaimWindowClose(currentHeight int64) int64 {
-	if currentHeight >= w.ClaimWindowClose {
-		return 0
-	}
-	return w.ClaimWindowClose - currentHeight
-}
-
-// BlocksUntilProofWindowClose returns blocks remaining until proof window closes.
-func (w SessionWindow) BlocksUntilProofWindowClose(currentHeight int64) int64 {
-	if currentHeight >= w.ProofWindowClose {
-		return 0
-	}
-	return w.ProofWindowClose - currentHeight
-}
